feat(build): Windows portability + libc-free net_runtime/io_engine
* build-windows.ps1: PowerShell entrypoint that loads MSVC, ensures
  ninja, builds local LLVM, then runs cargo with llvm/install/bin on
  PATH (lld-link is needed from the first crate, so LLVM must be ready
  before cargo starts — build.rs's own auto-build runs too late).
* llvm/build.sh: detect MINGW/MSYS/Cygwin, find .exe-suffixed
  llvm-config, validate partial clones, retry transient git failures.
* verum_mlir: target the bindgen MlirDiagnosticSeverity / MlirWalkOrder
  / MlirWalkResult aliases — bindgen emits u32 on Linux/macOS but i32
  on MSVC, so hard-coding either width fails to compile cross-platform.
* verum_cli/hooks: gate PermissionsExt + chmod under cfg(unix); Git for
  Windows runs hooks through bash regardless of permission bits.
* verum_error/Cargo.toml: move serde/serde_json out of the cfg(unix)
  block (they were accidentally Unix-only and broke Windows builds).
* verum_vbc/net_runtime: replace libc recv/send/recvfrom/accept4/
  getsockopt + sockaddr_storage decoding with std::net TcpStream /
  UdpSocket. Drop verify_real_eof, platform_msg_nosignal, and
  decode_sockaddr_storage helpers — std::net surfaces EOF and
  BrokenPipe as typed errors, set_nonblocking handles the readiness
  loop. Local EINVAL=22 keeps validation paths libc-free.
* verum_vbc/io_engine: close kqueue/epoll fd via OwnedFd::from_raw_fd
  drop instead of libc::close (cfg(unix) gated).
* verum_vbc/handlers: remove dead ctx_runtime module declaration.

+++ crates/llvm/verum_mlir/src/ir/operation/operation_like.rs
@@ -26,17 +26,18 @@ use super::{
 /// Order in which to traverse an operation tree.
 //
 
-// NOTE: `MlirWalkOrder_*` constants in the regenerated `verum_mlir_sys`
-// bindings (LLVM 21) are `u32`, not the historical `i32`. The enum repr
-// MUST match, otherwise the discriminant initializers `= Mlir..._*` fail
-// with E0308. The same applies to `WalkResult` below.
+// NOTE: bindgen types the `MlirWalkOrder_*` / `MlirWalkResult_*` constants
+// per the target ABI — `u32` on Linux/macOS, `i32` on Windows/MSVC (plain
+// C enums are signed there). A `#[repr(u32)]` enum needs `u32` discriminant
+// initializers, so each constant is cast with `as u32`; the values are
+// small and non-negative, so the cast is exact on every platform.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[repr(u32)]
 pub enum WalkOrder {
     /// Visit the operation before its nested regions.
-    PreOrder = MlirWalkOrder_MlirWalkPreOrder,
+    PreOrder = MlirWalkOrder_MlirWalkPreOrder as u32,
     /// Visit the operation after its nested regions.
-    PostOrder = MlirWalkOrder_MlirWalkPostOrder,
+    PostOrder = MlirWalkOrder_MlirWalkPostOrder as u32,
 }
 
 /// Control flow action returned by the walk callback.
@@ -44,11 +45,11 @@ pub enum WalkOrder {
 #[repr(u32)]
 pub enum WalkResult {
     /// Continue into this operation’s children.
-    Advance = MlirWalkResult_MlirWalkResultAdvance,
+    Advance = MlirWalkResult_MlirWalkResultAdvance as u32,
     /// Terminate the entire walk immediately.
-    Interrupt = MlirWalkResult_MlirWalkResultInterrupt,
+    Interrupt = MlirWalkResult_MlirWalkResultInterrupt as u32,
     /// Don’t visit this operation’s children, but keep walking siblings.
-    Skip = MlirWalkResult_MlirWalkResultSkip,
+    Skip = MlirWalkResult_MlirWalkResultSkip as u32,
 }
 
 pub trait OperationLike<'c: 'a, 'a>: Display + 'a {
